version 1.0 25-12-2019 - added into IndexRestController GET method by fire Id and iteration number. - added database layer with integration Spring JPA and supporting two individual DataSources: geometries(OSM data) and fires. - deleted redundant files from db and core package. - change logic for searching intersections agent ways with polygons in NetworkUtils.findIntersectionPoint.
In the next update will be implemented relief constraint for agents, new datasource for relief data, support external configuration application.properties

+++ src/main/java/ru/nachos/core/replanning/handlers/AgentTooCloseMovedHandlerImpl.java
@@ -27,7 +27,7 @@ public class AgentTooCloseMovedHandlerImpl implements AgentTooCloseMovedHandler 
         while (iterator.hasNext()){
             source = iterator.next();
             neighbour = source.getRightNeighbour();
-            if (source.getCoordinate().distance(neighbour.getCoordinate()) < info.getAgentDistance()/3){
+            if (source.getCoordinate().distance(neighbour.getCoordinate()) < (double) info.getAgentDistance()/3){
                 removeCandidates.add(source.getId());
             }
         }
